add import pos machine record function

+++ src/EasyERP.Web/Controllers/EmployeeController.cs
@@ -1,4 +1,7 @@
-﻿namespace EasyERP.Web.Controllers
+﻿using Doamin.Service.Security;
+using EasyErp.Core;
+
+namespace EasyERP.Web.Controllers
 {
     using System;
     using System.Collections.Generic;
@@ -11,38 +14,61 @@
     using EasyERP.Web.Models.Employee;
     using Infrastructure;
 
-    public class EmployeeController : Controller
+    public class EmployeeController :  BaseAdminController
     {
         private readonly IEmployeeService employeeService;
-
+        private readonly IPermissionService permissionService;
+        private readonly IWorkContext workContext;
         private readonly ITimesheetService<WorkTimeStatistic> timesheetService;
 
         public EmployeeController(
             IEmployeeService employeeService,
-            ITimesheetService<WorkTimeStatistic> timesheetService)
+            IPermissionService permissionService,
+            ITimesheetService<WorkTimeStatistic> timesheetService,
+            IWorkContext workContext)
         {
             this.employeeService = employeeService;
             this.timesheetService = timesheetService;
+            this.permissionService = permissionService;
+            this.workContext = workContext;
         }
 
         // GET: /Employee/
         public ActionResult Index()
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+                
             return View();
         }
 
         public ActionResult Create()
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
             return View();
         }
 
         public ActionResult Timesheet()
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
             return View();
         }
 
         public ActionResult Edit(int id)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
             var e = employeeService.GetEmployeeById(id);
             if (e != null)
             {
@@ -55,20 +81,32 @@
         [HttpPost]
         public ActionResult Edit(EmployeeModel model)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
             var e = Mapper.Map<EmployeeModel, Employee>(model);
             if (e != null)
             {
                 employeeService.UpdateEmployee(e);
             }
+
             return RedirectToAction("Index");
         }
 
         [HttpPost]
         public ActionResult Create(EmployeeModel employee)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
             var e = Mapper.Map<EmployeeModel, Employee>(employee);
             if (e != null)
             {
+                e.StoreId = workContext.CurrentUser.StoreId;
                 employeeService.AddEmployee(e);
             }
 
@@ -78,6 +116,11 @@
         [HttpPost]
         public JsonResult Delete(List<int> ids)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             if (ids != null)
             {
                 employeeService.DeleteEmployeeByIds(ids);
@@ -87,6 +130,11 @@
 
         public JsonResult EmployeeList(int skip, int take, int page, int pageSize)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             var employees = employeeService.GetEmployees(page, pageSize);
             if (employees != null)
             {
@@ -112,6 +160,11 @@
 
         public JsonResult GetTimeSheetByDate(string date, int page, int pageSize)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             DateTime selectedDate;
 
             if (!DateTime.TryParse(date, out selectedDate))
@@ -141,6 +194,11 @@
         [HttpPost]
         public JsonResult UpdateTimesheet(TimesheetModel model)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             DateTime selectedDate;
 
             if (!DateTime.TryParse(model.DateOfWeek, out selectedDate))
