+++ src/EasyERP.Web/Controllers/StoreSaleController.cs
@@ -1,4 +1,13 @@
-﻿namespace EasyERP.Web.Controllers
+﻿using System.Globalization;
+using System.IO;
+using System.Linq;
+using System.Web;
+using System.Xml.Linq;
+using Doamin.Service.Security;
+using Doamin.Service.Stores;
+using Domain.Model.Stores;
+
+namespace EasyERP.Web.Controllers
 {
     using System;
     using System.Collections.Generic;
@@ -8,9 +17,10 @@
     using Doamin.Service.StoreSale;
     using Domain.Model.Orders;
     using Domain.Model.Payments;
+    using EasyErp.Core;
     using EasyERP.Web.Models.StoreSale;
 
-    public class StoreSaleController : Controller
+    public class StoreSaleController : BaseAdminController 
     {
         private readonly ICustomerService customerService;
 
@@ -18,48 +28,90 @@
 
         private readonly IStoreSaleService storeSaleService;
 
+        private readonly IPostRetailService retailService;
+
+        private readonly IPermissionService permissionService;
+
+        private readonly IWorkContext workContext;
+
         public StoreSaleController(
+            IPostRetailService retailService,
             IStoreSaleService storeSaleService,
             IProductService productService,
-            ICustomerService customerService)
+            ICustomerService customerService,
+            IWorkContext workContext,
+            IPermissionService permissionService)
         {
             this.storeSaleService = storeSaleService;
             this.productService = productService;
             this.customerService = customerService;
+            this.retailService = retailService;
+            this.permissionService = permissionService;
+            this.workContext = workContext;
         }
 
         // GET: /StoreSales/
         public ActionResult Index()
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
             return View();
         }
 
         public ActionResult Create()
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
             return View();
         }
 
         [HttpPost]
         public JsonResult AddOrderItem(OrderItemModel orderItem)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             return Json(orderItem);
         }
 
         [HttpPost]
         public JsonResult DeleteOrderItem(OrderItemModel orderItem)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             return Json(orderItem);
         }
 
         [HttpPost]
         public JsonResult UpdateOrderItem(OrderItemModel orderItem)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             return Json(orderItem);
         }
 
         [HttpGet]
         public JsonResult AutoCompleteCustomers(string name)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             var customers = customerService.GetCustomersByName(name);
 
             var jsons = new List<object>();
@@ -81,6 +133,11 @@
         [HttpGet]
         public JsonResult AutoCompleteProducts(string name)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             var products = productService.GetAutoCompleteProducts(name);
 
             var jsons = new List<object>();
@@ -101,18 +158,34 @@
 
         public ActionResult Retail()
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
+            if (this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
             return View();
         }
 
         [HttpPost]
         public JsonResult Create(OrderModel model)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             var order = new Order
             {
                 CustomerId = model.CustomerId,
                 Name = model.Title,
                 CreatedOnUtc = DateTime.Now,
                 OrderGuid = Guid.NewGuid(),
+                StoreId = workContext.CurrentUser.StoreId,
                 Payment = new Payment()
             };
 
@@ -153,6 +226,11 @@
 
         public JsonResult OrderList(int skip, int take, int page, int pageSize)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             var orders = storeSaleService.GetOrders(page, pageSize);
             if (orders != null)
             {
@@ -183,14 +261,95 @@
             return Json(new {}, JsonRequestBehavior.AllowGet);
         }
 
-        [HttpPost]
+       [HttpPost]
         public JsonResult Delete(List<int> ids)
         {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
             if (ids != null)
             {
                 this.storeSaleService.DeleteOrderByIds(ids);
             }
             return Json(ids);
+        }
+
+        [HttpPost]
+        public ActionResult UploadRetail()
+        {
+            if (this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedView();
+            }
+
+            HttpPostedFileBase file = Request.Files["retialfile"];
+            if (file != null)
+            {
+                Stream fileStream = file.InputStream;
+                XDocument doc = XDocument.Load(fileStream);
+                var nodes = doc.Descendants("sale_item");
+                List<PostRetail> retails = new List<PostRetail>();
+                foreach (var node in nodes)
+                {
+                    PostRetail retail = new PostRetail();
+                    string itemNo = node.Elements("item_number").Select(t => (string) t).FirstOrDefault();
+
+                    var product = this.productService.GetProductByItemNo(itemNo);
+
+                    string date = node.Elements("date").Select(t => (string) t).FirstOrDefault();
+                    retail.ProductId = product.Id;
+                    retail.StoreId = 0;// this need to update according the user
+                    retail.Date = DateTime.ParseExact(date, "yyyyMMdd", new CultureInfo("zh-CN", true));
+                    retail.Quantity = node.Elements("sale_quantity").Select(t => (double)t).FirstOrDefault();
+                    retail.Price = node.Elements("sale_price").Select(t => (decimal)t).FirstOrDefault();
+                    retail.Cost = node.Elements("cost_price").Select(t => (decimal)t).FirstOrDefault();
+                    retails.Add(retail);
+                }
+                this.retailService.AddRecords(retails);
+
+                return RedirectToAction("Retail");
+            }
+
+            return Json(null);
+        }
+
+        [HttpGet]
+        public JsonResult RetailList(int skip, int take, int page, int pageSize)
+        {
+            if (!this.permissionService.Authorize(StandardPermissionProvider.ManageStores))
+            {
+                return AccessDeniedJson();
+            }
+
+            var retials = this.retailService.ShowPostRecord(page, pageSize);
+            if (retials != null)
+            {
+                List<object> listModel = new List<object>();
+                foreach (var retial in retials)
+                {
+                    object o = new
+                    {
+                        ProductName = retial.Product.Name,
+                        Price = retial.Price,
+                        //Cost = retial.Cost,
+                        Quantity = retial.Quantity,
+                        Date = retial.Date.ToString(CultureInfo.InvariantCulture),
+                        TotalAmount = (decimal) retial.Quantity*retial.Price
+                    };
+
+                    listModel.Add(o);
+                }
+                return Json(
+                         new
+                         {
+                             total = retials.TotalRecords,
+                             data = listModel
+                         },
+                         JsonRequestBehavior.AllowGet);
+            }
+            return Json(new { }, JsonRequestBehavior.AllowGet);
         }
     }
 }
